Move stateless methods to new modules in lib directory
