+++ app/models/ingredient.rb
@@ -1,22 +1,11 @@
 class Ingredient < ApplicationRecord
+  include IngredientCategories
+
   has_many :recipe_ingredients
   has_many :recipes, through: :recipe_ingredients
   validates :name, uniqueness: true
 
   before_save :assign_category
-
-  SWEETENERS = %w[
-    sugar brown\ sugar corn\ syrup
-    agave molasses honey maple\ syrup stevia
-  ].freeze
-  FATS = %w[butter cream sour\ cream canola\ oil olive\ oil margerine].freeze
-  WATER = %w[water milk].freeze
-  FLOURS = %w[
-    flour bread\ flour high\ gluten\ flour ap\ flour all\ purpose\ flour
-    spelt wheat\ flour whole\ wheat\ flour cake\ flour pastry\ flour semolina
-    durum corn\ meal flax\ meal
-  ].freeze
-
 
   def self.create_list(list)
     list.map { |name| Ingredient.find_or_create_by(name: name) }
