+++ app/models/recipe.rb
@@ -1,4 +1,6 @@
 class Recipe < ApplicationRecord
+  include RecipeFamilyInfo
+
   belongs_to :user
   has_many :recipe_ingredients, dependent: :destroy
   has_many :ingredients, through: :recipe_ingredients
@@ -40,22 +42,6 @@ class Recipe < ApplicationRecord
   def total_percent
     recipe_ingredients.reduce(0) { |sum, recipe_ingredient| sum + recipe_ingredient.bakers_percentage }.round(2)
   end
-
-  def sweetener_percentage
-    sweets = sweetener_amounts
-    calculate_percentage(sweets)
-  end
-
-  def fat_percentage
-    fats = fat_amounts
-    calculate_percentage(fats)
-  end
-
-  def water_percentage
-    water = water_amt
-    calculate_percentage(water)
-  end
-
   def ingredient_list
     list = {}
     recipe_ingredients.includes(:ingredient).each do |recipe_ingredient|
@@ -79,66 +65,6 @@ class Recipe < ApplicationRecord
     when rich  then self[:family] = 'Rich'
     when slack then self[:family] = 'Slack'
     end
-  end
-
-  def lean
-    return true if sweet_and_fat_amts.all? { |amt| low.include?(amt) }
-  end
-
-  def soft
-    if (water_percentage + fat_percentage) < 70.0 &&
-        moderate.include?(sweetener_percentage) &&
-        moderate.include?(fat_percentage)
-      true
-    end
-  end
-
-  def rich
-    if (moderate.include?(sweetener_percentage) &&
-        high.include?(fat_percentage)) ||
-        high.include?(fat_percentage)
-      true
-    end
-  end
-
-  def slack
-    return true if water_percentage + fat_percentage > 70.0
-  end
-
-  def sweet
-    return true if sweet_and_fat_amts.all? { |amt| high.include?(amt) }
-  end
-
-  def low
-    (0.0..4.99)
-  end
-
-  def moderate
-    (5.0..10.0)
-  end
-
-  def high
-    (11.0..25.0)
-  end
-
-  def sweet_and_fat_amts
-    [sweetener_percentage, fat_percentage]
-  end
-
-  def calculate_percentage(category)
-    ((category / flour_amts) * 100).round(2)
-  end
-
-  def sweetener_amounts
-    sum_recipe_ingredient_amounts('sweetener')
-  end
-
-  def fat_amounts
-    sum_recipe_ingredient_amounts('fat')
-  end
-
-  def water_amt
-    sum_recipe_ingredient_amounts('water')
   end
 
   def sum_recipe_ingredient_amounts(category)
