+++ lib/ingredient_categories.rb
@@ -1,0 +1,18 @@
+module IngredientCategories
+
+  SWEETENERS = %w[
+    sugar brown\ sugar corn\ syrup
+    agave molasses honey maple\ syrup stevia
+  ].freeze
+
+  FATS = %w[butter cream sour\ cream canola\ oil olive\ oil margerine].freeze
+
+  WATER = %w[water milk].freeze
+
+  FLOURS = %w[
+    flour bread\ flour high\ gluten\ flour ap\ flour all\ purpose\ flour
+    spelt wheat\ flour whole\ wheat\ flour cake\ flour pastry\ flour semolina
+    durum corn\ meal flax\ meal
+  ].freeze
+end
+
